Split all assets into their own modules
Old code had the main index file cluttered with a large amount of
classes that needed to be split out into their own modules for
readability and removal of cyclic dependencies.
