+++ lib/asset/verifier.js
@@ -3,7 +3,7 @@
 const fileType = require('file-type');
 const Validator = require('validator');
 
-const { logger } = require('./logging');
+const { logger } = require('../logging');
 
 const MAX_TITLE_LENGTH = 140;
