+++ lib/errors.js
@@ -1,0 +1,8 @@
+'use strict';
+
+class ValidationError extends Error {
+}
+
+module.exports = {
+    ValidationError,
+};
