+++ lib/util.js
@@ -4,15 +4,16 @@
 
 const cheerio = require('cheerio');
 const crypto = require('crypto');
-const feedparser = require('feedparser');
+const Feedparser = require('feedparser');
 const parseDataUrl = require('parse-data-url');
 const request = require('request');
 const url = require('url');
 
 const config = require('./config');
-const libingester = require('./index');
 const { logger } = require('./logging');
 const somaDOM = require('./soma_dom');
+const ImageAsset = require('./asset/imageAsset');
+const VideoAsset = require('./asset/videoAsset');
 
 const USER_AGENT = 'libingester';
 
@@ -64,6 +65,29 @@ function delayExecution(timeout) {
     return new Promise(resolve => {
         setTimeout(resolve, timeout);
     });
+}
+
+function encodeUri(uri) {
+    // Node's URL parser normalizes the URL and parses e.g.
+    // IDN hostnames into their Punycode representation.
+    const parsed = url.format(url.parse(uri));
+
+    // Go through and escape the URI.
+    return parsed.split('').map(chr => {
+        if (SAFE_ENCODING_CHARS.indexOf(chr) >= 0) {
+            return chr;
+        }
+
+        // Encode the code-point into UTF-8.
+        const buffer = Buffer.from(chr, 'utf8');
+
+        let encodedString = '';
+        buffer.forEach(n => {
+            encodedString += `%${n.toString(16).toUpperCase()}`;
+        });
+
+        return encodedString;
+    }).join('');
 }
 
 function fetchWithRetry(options, attempt = 1) {
@@ -141,7 +165,7 @@ function getDocBaseUri(doc, baseUri) {
  * @memberof util
  */
 function getEmbeddedVideoAsset(videoTag, videoUri) {
-    const asset = new libingester.VideoAsset();
+    const asset = new VideoAsset();
     asset.set_download_uri(videoUri);
     asset.set_last_modified_date(new Date());
 
@@ -193,7 +217,7 @@ function getImageSource($img, baseUri) {
 function downloadImage(uri) {
     const encodedUri = encodeUri(uri);
 
-    const asset = new libingester.ImageAsset();
+    const asset = new ImageAsset();
     asset.set_canonical_uri(encodedUri);
     asset.set_last_modified_date(new Date());
 
@@ -286,7 +310,7 @@ function downloadImageFromImgTag($img, baseUri) {
         const contentHash = getSha256(binaryData);
         const canonicalUri = `data:${mediaType};uri=${baseUri};sha256=${contentHash};`;
 
-        const asset = new libingester.ImageAsset();
+        const asset = new ImageAsset();
         asset.set_last_modified_date(new Date());
         asset.set_canonical_uri(encodeUri(canonicalUri));
         asset.set_image_data(mediaType, binaryData);
@@ -310,31 +334,6 @@ function downloadImageFromImgTag($img, baseUri) {
 
     return asset;
 }
-
-
-function encodeUri(uri) {
-    // Node's URL parser normalizes the URL and parses e.g.
-    // IDN hostnames into their Punycode representation.
-    const parsed = url.format(url.parse(uri));
-
-    // Go through and escape the URI.
-    return parsed.split('').map(chr => {
-        if (SAFE_ENCODING_CHARS.indexOf(chr) >= 0) {
-            return chr;
-        }
-
-        // Encode the code-point into UTF-8.
-        const buffer = Buffer.from(chr, 'utf8');
-
-        let encodedString = '';
-        buffer.forEach(n => {
-            encodedString += `%${n.toString(16).toUpperCase()}`;
-        });
-
-        return encodedString;
-    }).join('');
-}
-
 
 /**
  * Get parsed settings to limit the ingested items.
@@ -363,96 +362,6 @@ function getIngestionLimits(maxItems = Infinity, maxDaysOld = 1) {
     return {
         max_items: maxItems,
         oldest_date: oldestDate,
-    };
-}
-
-/**
- * A utility for fetching entries from an RSS feed for further processing.
- *
- * The `max_items` and `max_days_old` parameters can be overridden by the
- * `LIBINGESTER_MAX_ITEMS` and `LIBINGESTER_MAX_DAYS_OLD` environment
- * variables, respectively.
- *
- * @param {string|Array|Function} feed - URL of RSS feed to read, or an array of
- *   URLs of RSS feeds to read, or a paginator function that returns a URL given
- *   a page number
- * @param {number} [max_items=Infinity] - Maximum number of entries to return
- * @param {number} [max_days_old=1] - Maximum age of entries returned, in days
- * @returns {Promise<Object[]>} - Array of RSS feed entries
- * @memberof util
- */
-function fetchRssEntries(feed, max_items = Infinity, max_days_old = 1) {
-    // Allow environment variables to override this
-    const limits = getIngestionLimits(max_items, max_days_old);
-
-    return _fetchRssPage(feed, 1, [], new Set(), limits.max_items,
-                         limits.oldest_date);
-}
-
-function _fetchRssJson(feedUrl) {
-    return new Promise((resolve, reject) => {
-        const req = request.get(feedUrl);
-        const parser = new feedparser();
-
-        req.on('error', reject);
-        parser.on('error', reject);
-
-        const feedData = { items: [] };
-
-        parser.on('meta', meta => {
-            feedData.meta = meta;
-        });
-
-        parser.on('data', article => {
-            feedData.items.push(article);
-        });
-
-        parser.on('end', () => resolve(feedData));
-
-        req.pipe(parser);
-    });
-}
-
-function _isWordpressFeed(feedJson) {
-    const hasGenerator = feedJson.meta.generator &&
-          feedJson.meta.generator.startsWith('https://wordpress.org/');
-    const hasXmlns = feedJson.meta['rss:site'] &&
-          feedJson.meta['rss:site']['@'] &&
-          feedJson.meta['rss:site']['@']['xmlns'] &&
-          feedJson.meta['rss:site']['@']['xmlns'].startsWith('com-wordpress');
-
-    return hasGenerator || hasXmlns;
-}
-
-function _fetchRssJsonWithRetry(feedUrl, attempt = 1) {
-    const maxRetries = parseInt(config.get_setting('max-retries'), 10);
-    const retryBackoffDelay = parseInt(config.get_setting('retry-backoff-delay'), 10);
-
-    return _fetchRssJson(feedUrl).catch(err => {
-        if (attempt > maxRetries) {
-            logger.warn(`Max retries (${maxRetries}) exceeded!`);
-            throw err;
-        }
-
-        const shouldRetry = (err.message.match('Unexpected end') ||
-                             err.message.match('Not a feed'));
-        if (shouldRetry) {
-            const timeout = Math.pow(2, attempt) * retryBackoffDelay;
-            logger.debug(`Delaying execution of ${feedUrl} by ${timeout}`);
-
-            return delayExecution(timeout).then(() => {
-                return _fetchRssJsonWithRetry(feedUrl, attempt + 1);
-            });
-        }
-
-        // We reach here if unexpected error
-        throw err;
-    });
-}
-
-function _doCreateWordpress(feed) {
-    return function (pageNum) {
-        return `${feed}?paged=${pageNum}`;
     };
 }
 
@@ -533,6 +442,96 @@ function _fetchRssPage(feed, page, items, allLinks, maxItems, oldestDate) {
 }
 
 /**
+ * A utility for fetching entries from an RSS feed for further processing.
+ *
+ * The `max_items` and `max_days_old` parameters can be overridden by the
+ * `LIBINGESTER_MAX_ITEMS` and `LIBINGESTER_MAX_DAYS_OLD` environment
+ * variables, respectively.
+ *
+ * @param {string|Array|Function} feed - URL of RSS feed to read, or an array of
+ *   URLs of RSS feeds to read, or a paginator function that returns a URL given
+ *   a page number
+ * @param {number} [max_items=Infinity] - Maximum number of entries to return
+ * @param {number} [max_days_old=1] - Maximum age of entries returned, in days
+ * @returns {Promise<Object[]>} - Array of RSS feed entries
+ * @memberof util
+ */
+function fetchRssEntries(feed, max_items = Infinity, max_days_old = 1) {
+    // Allow environment variables to override this
+    const limits = getIngestionLimits(max_items, max_days_old);
+
+    return _fetchRssPage(feed, 1, [], new Set(), limits.max_items,
+                         limits.oldest_date);
+}
+
+function _fetchRssJson(feedUrl) {
+    return new Promise((resolve, reject) => {
+        const req = request.get(feedUrl);
+        const parser = new Feedparser();
+
+        req.on('error', reject);
+        parser.on('error', reject);
+
+        const feedData = { items: [] };
+
+        parser.on('meta', meta => {
+            feedData.meta = meta;
+        });
+
+        parser.on('data', article => {
+            feedData.items.push(article);
+        });
+
+        parser.on('end', () => resolve(feedData));
+
+        req.pipe(parser);
+    });
+}
+
+function _isWordpressFeed(feedJson) {
+    const hasGenerator = feedJson.meta.generator &&
+          feedJson.meta.generator.startsWith('https://wordpress.org/');
+    const hasXmlns = feedJson.meta['rss:site'] &&
+          feedJson.meta['rss:site']['@'] &&
+          feedJson.meta['rss:site']['@']['xmlns'] &&
+          feedJson.meta['rss:site']['@']['xmlns'].startsWith('com-wordpress');
+
+    return hasGenerator || hasXmlns;
+}
+
+function _fetchRssJsonWithRetry(feedUrl, attempt = 1) {
+    const maxRetries = parseInt(config.get_setting('max-retries'), 10);
+    const retryBackoffDelay = parseInt(config.get_setting('retry-backoff-delay'), 10);
+
+    return _fetchRssJson(feedUrl).catch(err => {
+        if (attempt > maxRetries) {
+            logger.warn(`Max retries (${maxRetries}) exceeded!`);
+            throw err;
+        }
+
+        const shouldRetry = (err.message.match('Unexpected end') ||
+                             err.message.match('Not a feed'));
+        if (shouldRetry) {
+            const timeout = Math.pow(2, attempt) * retryBackoffDelay;
+            logger.debug(`Delaying execution of ${feedUrl} by ${timeout}`);
+
+            return delayExecution(timeout).then(() => {
+                return _fetchRssJsonWithRetry(feedUrl, attempt + 1);
+            });
+        }
+
+        // We reach here if unexpected error
+        throw err;
+    });
+}
+
+function _doCreateWordpress(feed) {
+    return function (pageNum) {
+        return `${feed}?paged=${pageNum}`;
+    };
+}
+
+/**
  * A utility to create a pagination function for Wordpress-generated feeds,
  * since they are so common. Make sure to verify that your feed is generated by
  * Wordpress before using this!
@@ -562,4 +561,3 @@ module.exports = {
     get_ingestion_limits: getIngestionLimits,
     set_user_agent: setUserAgent,
 };
-
