+++ test/lib/asset/verifier.test.js
@@ -2,7 +2,7 @@
 
 const expect = require('chai').expect;
 
-const verify = require('../../lib/verify');
+const verifier = require('../../../lib/asset/verifier');
 
 const basic_metadata = {
     "assetID": "03d9d97e112b3b4513d246bda42c853656f200e8",
@@ -18,22 +18,22 @@ const basic_metadata = {
     "cdnFilename": "03d9d97e112b3b4513d246bda42c853656f200e8.data",
 };
 
-describe('verify', function() {
+describe('verifier', function() {
     it('passes real metadata correctly', function() {
-        expect(() => verify.verify_metadata(basic_metadata)).not.to.throw();
+        expect(() => verifier.verify_metadata(basic_metadata)).not.to.throw();
     });
 
     it('fails fake metadata', function() {
-        expect(() => verify.verify_metadata({})).to.throw();
+        expect(() => verifier.verify_metadata({})).to.throw();
     });
 
     it('tests asset IDs correctly', function() {
         const fake_id_metadata = Object.assign({}, basic_metadata, { "assetID": "12345" });
-        expect(() => verify.verify_metadata(fake_id_metadata)).to.throw();
+        expect(() => verifier.verify_metadata(fake_id_metadata)).to.throw();
     });
 
     it('tests tags correctly', function() {
         const bad_tags_metadata = Object.assign({}, basic_metadata, { "tags": [ ["A", "B"] ] });
-        expect(() => verify.verify_metadata(fake_id_metadata)).to.throw();
+        expect(() => verifier.verify_metadata(fake_id_metadata)).to.throw();
     });
 });
