+++ test/lib/index.test.js
@@ -4,16 +4,31 @@ const expect = require('chai').expect;
 const fs = require('fs');
 const proxyquire = require('proxyquire');
 
+
+const BaseAsset = require('../../lib/asset/baseAsset');
 const config = require('../../lib/config');
 
-const libingester = proxyquire('../../lib/index', {
-    './verify': {
-        verify_metadata: () => {},
-        verify_manifest_entry: () => {},
+const verifier = require('../../lib/asset/verifier');
+
+const mockVerifier = {
+    verify_metadata: () => {
+        return null;
     },
+    verify_manifest_entry: () => {
+        return null;
+    },
+    VerificationError: verifier.VerificationError,
+};
+
+const libingester = proxyquire('../../lib/index', {
+    './asset/verifier': mockVerifier,
 });
 
-class MockAsset extends libingester.BaseAsset {
+class MockAsset extends BaseAsset {
+    constructor() {
+        super(mockVerifier);
+    }
+
     fails_with_error(err) {
         this._err = err;
     }
@@ -368,189 +383,5 @@ describe('MockAsset', () => {
         expect(() => {
             asset.set_synopsis(undefined);
         }).to.not.throw();
-    });
-});
-
-describe('ImageAsset', () => {
-    it('can serialize out correctly', () => {
-        const asset = new libingester.ImageAsset();
-        const thumbnailAsset = new libingester.ImageAsset();
-        asset.set_title('Test Asset');
-        asset.set_synopsis('Test Asset synopsis');
-        asset.set_thumbnail(thumbnailAsset);
-        asset.set_license('Proprietary');
-        asset.set_canonical_uri('https://www.example.com/');
-        asset.set_last_modified_date(new Date(1492545280000));
-        asset.set_image_data('image/jpeg', 'asdf');
-
-        const metadata = asset.to_metadata();
-
-        // Check that asset ID and thumbnail asset ID are passed through
-        expect(metadata['assetID']).to.equal(asset.asset_id);
-        expect(metadata['thumbnail']).to.equal(thumbnailAsset.asset_id);
-        // Remove the ID fields before checking the rest
-        delete metadata['assetID'];
-        delete metadata['thumbnail'];
-
-        expect(metadata).to.deep.equal({
-            'objectType': 'ImageObject',
-            'contentType': 'image/jpeg',
-
-            'canonicalURI': 'https://www.example.com/',
-            'matchingLinks': ['https://www.example.com/'],
-
-            'title': 'Test Asset',
-            'synopsis': 'Test Asset synopsis',
-            'license': 'Proprietary',
-            'tags': [],
-            'lastModifiedDate': '2017-04-18T19:54:40.000Z',
-            'revisionTag': '2017-04-18T19:54:40.000Z',
-        });
-
-        const data = asset.to_data();
-        expect(data).to.equal('asdf');
-    });
-});
-
-describe('BlogArticle', () => {
-    let asset;
-
-    beforeEach(() => {
-        asset = new libingester.BlogArticle();
-        asset.set_title('Test Asset');
-        asset.set_license('Proprietary');
-        asset.set_canonical_uri('https://www.example.com/');
-        asset.set_last_modified_date(new Date(1492545280000));
-        asset.set_synopsis('a long time ago...');
-        asset.set_body('<h1>Word of the Day</h1>');
-        asset.set_author('Coco');
-        asset.set_date_published(new Date(1492545280000));
-        asset.set_read_more_text('More!');
-        asset.set_tags([ 'some', 'tags' ]);
-        asset.set_as_static_page();
-    });
-
-    it('can serialize out correctly', () => {
-        asset.render();
-
-        const metadata = asset.to_metadata();
-
-        delete metadata['assetID'];
-
-        expect(metadata['document']).to.contain('<h1>Word of the Day</h1>');
-        expect(metadata['document']).to.contain('More!');
-        // Match at least one CSS rule despite no custom SCSS
-        expect(metadata['document']).to.match(/<style(.|\n)*{(.|\n)*:(.|\n)*}(.|\n)*<\/style>/);
-        delete metadata['document'];
-
-        expect(metadata).to.deep.eql({
-            'objectType': 'ArticleObject',
-            'contentType': 'text/html',
-
-            'canonicalURI': 'https://www.example.com/',
-            'matchingLinks': ['https://www.example.com/'],
-
-            'title': 'Test Asset',
-            'license': 'Proprietary',
-            'tags': [ 'some', 'tags', 'EknStaticTag' ],
-            'synopsis': 'a long time ago...',
-            'lastModifiedDate': '2017-04-18T19:54:40.000Z',
-            'revisionTag': '2017-04-18T19:54:40.000Z',
-
-            'authors': ['Coco'],
-            'published': '2017-04-18T19:54:40.000Z',
-        });
-    });
-
-    it('renders the custom stylesheet', () => {
-        asset.set_custom_scss('@import "_default"; * { color:red; }');
-        asset.render();
-
-        const metadata = asset.to_metadata();
-        // Regex handles how libsass might minify the rendered CSS
-        expect(metadata['document']).to.match(/\*\s*{\s*color:\s*red;?\s*}/);
-    });
-    it('cleans newlines from synopsis', () => {
-        asset.set_synopsis('This is a line.\nThis is the same line.\n');
-        asset.render();
-        const metadata = asset.to_metadata();
-        expect(metadata['synopsis']).to.equal('This is a line. This is the same line.');
-    });
-});
-
-describe('NewsAsset', () => {
-    let asset;
-
-    beforeEach(() => {
-        asset = new libingester.NewsArticle();
-        asset.set_title('Test Asset');
-        asset.set_license('Proprietary');
-        asset.set_canonical_uri('https://www.example.com/');
-        asset.set_last_modified_date(new Date(1492545280000));
-        asset.set_body('<h1>Word of the Day</h1>');
-        asset.set_section('word_of_day');
-        asset.set_synopsis('a long time ago...');
-        asset.set_as_static_page();
-        asset.set_authors([ 'Merriam', 'Webster' ]);
-        asset.set_source('Dictionary');
-        asset.set_date_published(new Date(1492545280000));
-        asset.set_read_more_link('More!');
-        asset.set_lede('<p>Exciting paragraph</p>');
-    });
-
-    it('can serialize out correctly', () => {
-        asset.render();
-
-        const metadata = asset.to_metadata();
-
-        // Remove randomness -- should probably be a mock if I can
-        // figure out how to use it.
-        delete metadata['assetID'];
-
-        expect(metadata['document']).to.contain('<h1>Word of the Day</h1>');
-        expect(metadata['document']).to.contain('More!');
-        expect(metadata['document']).to.contain('<p>Exciting paragraph</p>');
-        delete metadata['document'];
-
-        expect(metadata).to.deep.eql({
-            'objectType': 'ArticleObject',
-            'contentType': 'text/html',
-
-            'canonicalURI': 'https://www.example.com/',
-            'matchingLinks': ['https://www.example.com/'],
-
-            'title': 'Test Asset',
-            'license': 'Proprietary',
-            'tags': [ 'word_of_day', 'EknStaticTag' ],
-            'synopsis': 'a long time ago...',
-            'lastModifiedDate': '2017-04-18T19:54:40.000Z',
-            'revisionTag': '2017-04-18T19:54:40.000Z',
-
-            'authors': [ 'Merriam', 'Webster' ],
-            'sourceName': 'Dictionary',
-            'published': '2017-04-18T19:54:40.000Z',
-        });
-    });
-
-    it('renders the default stylesheet if no custom SCSS set', () => {
-        asset.render();
-
-        const metadata = asset.to_metadata();
-
-        // Match at least one CSS rule despite no custom SCSS
-        expect(metadata['document']).to.match(/<style(.|\n)*{(.|\n)*:(.|\n)*}(.|\n)*<\/style>/);
-    });
-
-    it('renders the custom SCSS', () => {
-        asset.set_custom_scss('@import "_default"; * { color:red; }');
-        asset.render();
-
-        const metadata = asset.to_metadata();
-        // Regex handles how libsass might minify the rendered CSS
-        expect(metadata['document']).to.match(/\*\s*{\s*color:\s*red;?\s*}/);
-    });
-
-    it('cannot use set_tags', () => {
-        expect(asset.set_tags).to.throw();
     });
 });
